refactor(analysis)!: Phase 6 S1 — relocate comparison + reconciliation engine -> analysis/
Create the analysis/ layer (above engine/) and git-mv the two cross-run runners
into it verbatim (behaviour-preserving, zero logic change):
- engine/comparison.py    -> analysis/comparison.py   (DualFrameworkRunner,
  CapitalImpactAnalyzer, TransitionalScheduleRunner)
- engine/reconciliation.py -> analysis/reconciliation.py (ReconciliationRunner)

analysis/ imports engine/contracts/rulebook/data downward (legal per the
pre-declared arch_check import-direction rules; analysis has no rule entry).
Repointed the 4 src consumers (api/service, api/reconciliation docstring,
ui/app/main, ui/views/reconciliation) + 11 test files, and removed the
engine/__init__ re-export of the comparison runners (engine may not import
analysis). Retired the now-dead arch_check allowlist entries keyed on the moved
paths (REGULATORY_SCALAR / VALIDATION_ENUM / REGIME_BOOL; checks 5/6/17 scan
engine/ only) and rebanked the ratchet (reconciliation's defensive sites left
engine/: fill_null 463->441, presence 367->353, collect_schema 164->154,
eager 38->33).

The reconciliation registry (RECONCILABLE_COMPONENTS) + recon config stay put
for now — S2 relocates them and resolves the contracts->registry layering knot.

Gate: ruff + ty clean; arch_check EXIT 0; full suite 7506 passed / 2 skipped.
