+++ scripts/arch_metrics.json
@@ -1,9 +1,9 @@
 {
   "_comment": "Architecture-debt ratchet baseline (arch_check check 11). Counts may not increase (cites_decorators may not decrease). Regenerate after an improvement with: python scripts/arch_check.py --update-baseline",
-  "engine_fill_null_sites": 463,
-  "engine_presence_guard_sites": 367,
-  "engine_collect_schema_sites": 164,
-  "engine_eager_collect_sites": 38,
+  "engine_fill_null_sites": 441,
+  "engine_presence_guard_sites": 353,
+  "engine_collect_schema_sites": 154,
+  "engine_eager_collect_sites": 33,
   "max_engine_module_loc": 1518,
   "cites_decorators": 187
 }
