+++ src/rwa_calc/ui/views/reconciliation.py
@@ -29,7 +29,7 @@ from typing import TYPE_CHECKING, SupportsFloat, cast
 
 import polars as pl
 
-from rwa_calc.engine.reconciliation import (
+from rwa_calc.analysis.reconciliation import (
     BUCKET_BREAK,
     BUCKET_EXACT,
     BUCKET_MISSING_LEFT,
