+++ tests/unit/engine/test_reconciliation.py
@@ -11,6 +11,14 @@ from __future__ import annotations
 import polars as pl
 import pytest
 
+from rwa_calc.analysis.reconciliation import (
+    BUCKET_BREAK,
+    BUCKET_EXACT,
+    BUCKET_MISSING_LEFT,
+    BUCKET_MISSING_RIGHT,
+    BUCKET_WITHIN,
+    ReconciliationRunner,
+)
 from rwa_calc.contracts.config import ComponentMapping, LegacyColumnMapping
 from rwa_calc.contracts.errors import (
     ERROR_RECON_DUPLICATE_LEGACY_KEY,
@@ -19,14 +27,6 @@ from rwa_calc.contracts.errors import (
     ERROR_RECON_LEGACY_COLUMN_MISSING,
 )
 from rwa_calc.contracts.protocols import ReconciliationRunnerProtocol
-from rwa_calc.engine.reconciliation import (
-    BUCKET_BREAK,
-    BUCKET_EXACT,
-    BUCKET_MISSING_LEFT,
-    BUCKET_MISSING_RIGHT,
-    BUCKET_WITHIN,
-    ReconciliationRunner,
-)
 
 
 def _ours() -> pl.LazyFrame:
